chore(STK-198): move and update sections banners
