--- packages/landing/components/QAndAAccordion/index.ts
@@ -1,1 +1,0 @@
-export * from "./QAndAAccordion";
