+++ packages/landing/components/index.ts
@@ -1,4 +1,2 @@
 export * from "./navbar";
-export * from "./QAndAAccordion";
 export * from "./sections";
-export * from "./TryStacklyBanner";
