+++ packages/landing/components/sections/HeroBanner/index.ts
@@ -1,0 +1,1 @@
+export * from "./HeroBanner";
