+++ packages/landing/components/sections/index.ts
@@ -1,1 +1,3 @@
-export * from "./MainSection";
+export * from "./FAQ";
+export * from "./HeroBanner";
+export * from "./TryStacklyBanner";
